修复 Copilot,golangci-lint 的问题
